Allow to update user's identity code and country code.
Previously, none of these values were allowed to be updated, but the
form had only identity_code disabled. This led to the idea that you
can change the country code, but not the identity code.

This is not correct, as both of these values are changeable during a
normal lifecycle (ie. when a person moves to another country)