Add DejaVu Sans Mono font
Unfortunately, there is no npm package that would provide it, so the
file needs to be included directly. Difference with Raleway font that
we use for Semantic UI framework is that it is done at a later stage
of compilation, so the file is included without problems in the final build.